==> V10/Jupyter_Notebooks_For_Taxi/1_Pre_Processing/1_Hands-on.ipynb <==
# Let's check the data on HDFS 

### File Sizes 

!hdfs dfs -du -s -h /taxi/raw/

### Looks like we have all years

!hdfs dfs -ls /taxi/raw/

### However, the data gets smaller at 2016.

!hdfs dfs -du -s -h /taxi/raw/*

### Let's see if the schema changes over time 

!hdfs dfs -ls -h /taxi/raw/2009/yellow_tripdata_2009-01.csv

!hdfs dfs -ls -h /taxi/raw/2017/yellow_tripdata_2017-01.csv

### Init Spark 

from pyspark.sql import SparkSession

spark = SparkSession \
    .builder \
    .appName("Hands On") \
    .getOrCreate()

df_2009_01 = spark.read.csv("/taxi/raw/2009/yellow_tripdata_2009-01.csv", header=True)

df_2017_01 = spark.read.csv("/taxi/raw/2017/yellow_tripdata_2017-01.csv", header=True)

###  df_2009_01 vs df_2017_01

There is a diff in the schema and we have to deal with it.
- Description of the [schema](https://www1.nyc.gov/assets/tlc/downloads/pdf/data_dictionary_trip_records_yellow.pdf)

df_2009_01.show(2)

df_2017_01.show(2)

### Count should be similar 

print(df_2009_01.count())
print(df_2017_01.count())

We have to analyse if less was recorded or if this is just a downward trend in general

### The Schema changed at 2016 - between June and July 

!hdfs dfs -du -s -h /taxi/raw/2016/*

## Column Names 

df_2009_01.printSchema()

df_2017_01.printSchema()

### Stopping Spark 

spark.stop()
