==> V10/Jupyter_Notebooks_For_Taxi/1_Pre_Processing/3_Pre_Processing_schema_changes.ipynb <==
!hdfs dfs -ls /taxi/raw_parquet/

from pyspark.sql import SparkSession

spark = SparkSession \
    .builder \
    .appName("Convert To Parquet and pick columns") \
    .getOrCreate()

# Schema 

As we have seen, the schema changed. Let us build a function which picks the common columns we need for future queries and rename them to the same column name 

Old:
```bash
+-----------+--------------------+---------------------+---------------+------------------+-------------------+------------------+---------+-----------------+-------------------+------------------+------------+------------------+---------+-------+-------+---------+------------------+
|vendor_name|Trip_Pickup_DateTime|Trip_Dropoff_DateTime|Passenger_Count|     Trip_Distance|          Start_Lon|         Start_Lat|Rate_Code|store_and_forward|            End_Lon|           End_Lat|Payment_Type|          Fare_Amt|surcharge|mta_tax|Tip_Amt|Tolls_Amt|         Total_Amt|
+-----------+--------------------+---------------------+---------------+------------------+-------------------+------------------+---------+-----------------+-------------------+------------------+------------+------------------+---------+-------+-------+---------+------------------+
|        VTS| 2009-01-04 02:52:00|  2009-01-04 03:02:00|              1|2.6299999999999999|-73.991956999999999|         40.721567|     null|             null|         -73.993803|40.695922000000003|        CASH|8.9000000000000004|      0.5|   null|      0|        0|9.4000000000000004|
|        VTS| 2009-01-04 03:31:00|  2009-01-04 03:38:00|              3|4.5499999999999998|-73.982101999999998|40.736289999999997|     null|             null|-73.955849999999998|40.768030000000003|      Credit|              12.1|      0.5|   null|      2|        0|              14.6|
+-----------+--------------------+---------------------+---------------+------------------+-------------------+------------------+---------+-----------------+-------------------+------------------+------------+------------------+---------+-------+-------+---------+------------------+
```

New:

```bash
+--------+--------------------+---------------------+---------------+-------------+----------+------------------+------------+------------+------------+-----------+-----+-------+----------+------------+---------------------+------------+
|VendorID|tpep_pickup_datetime|tpep_dropoff_datetime|passenger_count|trip_distance|RatecodeID|store_and_fwd_flag|PULocationID|DOLocationID|payment_type|fare_amount|extra|mta_tax|tip_amount|tolls_amount|improvement_surcharge|total_amount|
+--------+--------------------+---------------------+---------------+-------------+----------+------------------+------------+------------+------------+-----------+-----+-------+----------+------------+---------------------+------------+
|       1| 2017-01-09 11:13:28|  2017-01-09 11:25:45|              1|         3.30|         1|                 N|         263|         161|           1|       12.5|    0|    0.5|         2|           0|                  0.3|        15.3|
|       1| 2017-01-09 11:32:27|  2017-01-09 11:36:01|              1|          .90|         1|                 N|         186|         234|           1|          5|    0|    0.5|      1.45|           0|                  0.3|        7.25|
+--------+--------------------+---------------------+---------------+-------------+----------+------------------+------------+------------+------------+-----------+-----+-------+----------+------------+---------------------+------------+
```

We pick: 

```
Trip_Pickup_DateTime tpep_pickup_datetime
Trip_Dropoff_DateTime tpep_dropoff_datetime
Passenger_Count passenger_count
Trip_Distance trip_distance
Payment_Type payment_type
Tip_Amt tip_amount
Total_Amt total_amount
```

### Create Folder in HDFS

!hdfs dfs -mkdir /taxi/by_month/

### testing our approach first

year = "2009"
month = "01"

df = spark.read.parquet(f"/taxi/raw_parquet/{year}/{month}.parquet")\
    .withColumnRenamed("Trip_Pickup_DateTime","pickup_datetime")\
    .withColumnRenamed("Trip_Dropoff_DateTime","dropoff_datetime")\
    .withColumnRenamed("tpep_pickup_datetime","pickup_datetime")\
    .withColumnRenamed("tpep_dropoff_datetime","dropoff_datetime")\
    .withColumnRenamed("Passenger_Count","passenger_count")\
    .withColumnRenamed("Trip_Distance","trip_distance")\
    .withColumnRenamed("Payment_Type","payment_type")\
    .withColumnRenamed("Tip_Amt","tip_amount")\
    .withColumnRenamed("Total_Amt","total_amount")\
    .selectExpr(\
        "cast(pickup_datetime as timestamp)", \
        "cast(dropoff_datetime as timestamp)", \
        "cast(passenger_count as int)", \
        "cast(trip_distance as double)", \
        "cast(payment_type as string)", \
        "cast(tip_amount as double)", \
        "cast(total_amount as double)" \
    )

df.printSchema()

df.show(2)

year = "2021"
month = "03"

df = spark.read.parquet(f"/taxi/raw_parquet/{year}/{month}.parquet")\
    .withColumnRenamed("Trip_Pickup_DateTime","pickup_datetime")\
    .withColumnRenamed("Trip_Dropoff_DateTime","dropoff_datetime")\
    .withColumnRenamed("tpep_pickup_datetime","pickup_datetime")\
    .withColumnRenamed("tpep_dropoff_datetime","dropoff_datetime")\
    .withColumnRenamed("Passenger_Count","passenger_count")\
    .withColumnRenamed("Trip_Distance","trip_distance")\
    .withColumnRenamed("Payment_Type","payment_type")\
    .withColumnRenamed("Tip_Amt","tip_amount")\
    .withColumnRenamed("Total_Amt","total_amount")\
    .selectExpr(\
        "cast(pickup_datetime as timestamp)", \
        "cast(dropoff_datetime as timestamp)", \
        "cast(passenger_count as int)", \
        "cast(trip_distance as double)", \
        "cast(payment_type as string)", \
        "cast(tip_amount as double)", \
        "cast(total_amount as double)" \
    )

df.printSchema()

df.show(2)

## Read and Write

def read_and_write(year, month):
    df = spark.read.parquet(f"/taxi/raw_parquet/{year}/{month}.parquet")\
    .withColumnRenamed("Trip_Pickup_DateTime","pickup_datetime")\
    .withColumnRenamed("Trip_Dropoff_DateTime","dropoff_datetime")\
    .withColumnRenamed("tpep_pickup_datetime","pickup_datetime")\
    .withColumnRenamed("tpep_dropoff_datetime","dropoff_datetime")\
    .withColumnRenamed("Passenger_Count","passenger_count")\
    .withColumnRenamed("Trip_Distance","trip_distance")\
    .withColumnRenamed("Payment_Type","payment_type")\
    .withColumnRenamed("Tip_Amt","tip_amount")\
    .withColumnRenamed("Total_Amt","total_amount")\
    .selectExpr(\
        "cast(pickup_datetime as timestamp)", \
        "cast(dropoff_datetime as timestamp)", \
        "cast(passenger_count as int)", \
        "cast(trip_distance as double)", \
        "cast(payment_type as string)", \
        "cast(tip_amount as double)", \
        "cast(total_amount as double)" \
    )
    
    df.write.parquet(f"/taxi/by_month/{year}/{month}.parquet")

for year in ["2009", "2010", "2011", "2012", "2013", "2014", "2015", "2016", "2017", "2018", "2019", "2020", "2021", "2022"]:
    for month in ["01", "02", "03", "04", "05", "06", "07", "08", "09", "10", "11", "12"]:
        !echo processing {year}/{month}
        read_and_write(year, month)

## Stopping Spark 

spark.stop()
