==> V10/Jupyter_Notebooks_For_Taxi/1_Pre_Processing/4_Unifying_data_and_extra_cols.ipynb <==
from pyspark.sql import SparkSession

spark = SparkSession \
    .builder \
    .appName("Unify") \
    .getOrCreate()

# I want to take all data (not per month)

- adding the month and year as columns
- convert trip distance to km instead of miles
- adding a field `amount-tip`
- unifying payment_type

## Test Data 01/2022

year = "2022"
month = "01"
df = spark.read.parquet(f"/taxi/by_month/{year}/{month}.parquet")

### adding month/year as columns

You can use `input_file_name` to get the filename of the dataframe. Here we have the month/year available

from pyspark.sql.functions import input_file_name

df.withColumn("filename", input_file_name()).show(2, False)

from pyspark.sql.functions import input_file_name, split

df.withColumn("filename", split(input_file_name(), "/")).show(2, False)

from pyspark.sql.functions import input_file_name, split

df.withColumn("filename", split(input_file_name(), "/")).printSchema()

df\
.withColumn("filename", split(input_file_name(), "/"))\
.selectExpr("*", "element_at(filename, -3) as year", "substring_index(element_at(filename, -2), '.' , 1) as month")\
.drop("filename")\
.show(2, False)

### convert trip distance to km instead of miles

from pyspark.sql.functions import expr

df\
.withColumn("filename", split(input_file_name(), "/"))\
.selectExpr("*", "element_at(filename, -3) as year", "substring_index(element_at(filename, -2), '.' , 1) as month")\
.drop("filename")\
.withColumn("trip_distance", expr("trip_distance * 1.60934"))\
.show(2, False)

### adding a field `amount-tip`

df\
.withColumn("filename", split(input_file_name(), "/"))\
.selectExpr("*", "element_at(filename, -3) as year", "substring_index(element_at(filename, -2), '.' , 1) as month")\
.drop("filename")\
.withColumn("trip_distance", expr("trip_distance * 1.60934"))\
.withColumn("trip_amount", expr("total_amount-tip_amount"))\
.show(2, False)

### unifying payment_type

df\
.withColumn("filename", split(input_file_name(), "/"))\
.selectExpr("*", "element_at(filename, -3) as year", "substring_index(element_at(filename, -2), '.' , 1) as month")\
.drop("filename")\
.withColumn("trip_distance", expr("trip_distance * 1.60934"))\
.withColumn("trip_amount", expr("total_amount-tip_amount"))\
.select("payment_type")\
.distinct()\
.show(20, False)

# earlier data
old_df = spark.read.parquet(f"/taxi/by_month/2009/01.parquet")\
.withColumn("filename", split(input_file_name(), "/"))\
.selectExpr("*", "element_at(filename, -3) as year", "substring_index(element_at(filename, -2), '.' , 1) as month")\
.drop("filename")\
.withColumn("trip_distance", expr("trip_distance * 1.60934"))\
.withColumn("trip_amount", expr("total_amount-tip_amount"))


old_df.select("payment_type")\
.distinct()\
.show(20, False)

```
1 = Credit card 
2 = Cash
3 = No charge 
4 = Dispute
5 = Unknown 
6 = Voided trip
```

from pyspark.sql.functions import when, expr, col

old_df.withColumn('payment_type', \
                  when(col("payment_type") == "Credit", 1)\
                  .when(col("payment_type") == "CREDIT", 1)\
                  .when(col("payment_type") == "CASH", 2)\
                  .when(col("payment_type") == "Cash", 2)\
                  .when(col("payment_type") == "No Charge", 3)\
                  .when(col("payment_type") == "NO CHARGE", 3)\
                  .when(col("payment_type") == "Dispute", 4)\
                  .when(col("payment_type") == "DISPUTE", 4)\
                  .when(col("payment_type") == "Unknown", 5)\
                  .when(col("payment_type") == "UNKNOWN", 5)\
                  .when(col("payment_type") == "Voided Trip", 6)\
                  .when(col("payment_type") == "VOIDED TRIP", 6)\
                  .otherwise(expr("payment_type")))\
    .select("payment_type")\
    .distinct()\
    .show(20, False)


### All Together

df\
.withColumn("filename", split(input_file_name(), "/"))\
.selectExpr("*", "element_at(filename, -3) as year", "substring_index(element_at(filename, -2), '.' , 1) as month")\
.drop("filename")\
.withColumn("trip_distance", expr("trip_distance * 1.60934"))\
.withColumn("trip_amount", expr("total_amount-tip_amount"))\
.withColumn('payment_type', \
                  when(col("payment_type") == "Credit", 1)\
                  .when(col("payment_type") == "CREDIT", 1)\
                  .when(col("payment_type") == "CASH", 2)\
                  .when(col("payment_type") == "Cash", 2)\
                  .when(col("payment_type") == "No Charge", 3)\
                  .when(col("payment_type") == "NO CHARGE", 3)\
                  .when(col("payment_type") == "Dispute", 4)\
                  .when(col("payment_type") == "DISPUTE", 4)\
                  .when(col("payment_type") == "Unknown", 5)\
                  .when(col("payment_type") == "UNKNOWN", 5)\
                  .when(col("payment_type") == "Voided Trip", 6)\
                  .when(col("payment_type") == "VOIDED TRIP", 6)\
                  .otherwise(expr("payment_type")))\
.select(\
    "year", \
    "month", \
    "pickup_datetime", \
    "dropoff_datetime", \
    "passenger_count", \
    "trip_distance", \
    "payment_type", \
    "tip_amount", \
    "trip_amount", \
    "total_amount" \
).show(2, False)

## All Data

df = spark.read.parquet(f"/taxi/by_month/*/*.parquet")

all_data = df\
.withColumn("filename", split(input_file_name(), "/"))\
.selectExpr("*", "element_at(filename, -3) as year", "substring_index(element_at(filename, -2), '.' , 1) as month")\
.drop("filename")\
.withColumn("trip_distance", expr("trip_distance * 1.60934"))\
.withColumn("trip_amount", expr("total_amount-tip_amount"))\
.withColumn('payment_type', \
                  when(col("payment_type") == "Credit", 1)\
                  .when(col("payment_type") == "CREDIT", 1)\
                  .when(col("payment_type") == "CASH", 2)\
                  .when(col("payment_type") == "Cash", 2)\
                  .when(col("payment_type") == "No Charge", 3)\
                  .when(col("payment_type") == "NO CHARGE", 3)\
                  .when(col("payment_type") == "Dispute", 4)\
                  .when(col("payment_type") == "DISPUTE", 4)\
                  .when(col("payment_type") == "Unknown", 5)\
                  .when(col("payment_type") == "UNKNOWN", 5)\
                  .when(col("payment_type") == "Voided Trip", 6)\
                  .when(col("payment_type") == "VOIDED TRIP", 6)\
                  .otherwise(expr("payment_type")))\
.select(\
    "year", \
    "month", \
    "pickup_datetime", \
    "dropoff_datetime", \
    "passenger_count", \
    "trip_distance", \
    "payment_type", \
    "tip_amount", \
    "trip_amount", \
    "total_amount" \
)

### Payment Type is still bad... 

all_data = df\
.withColumn("filename", split(input_file_name(), "/"))\
.selectExpr("*", "element_at(filename, -3) as year", "substring_index(element_at(filename, -2), '.' , 1) as month")\
.drop("filename")\
.withColumn("trip_distance", expr("trip_distance * 1.60934"))\
.withColumn("trip_amount", expr("total_amount-tip_amount"))\
.withColumn('payment_type', \
                  when(col("payment_type") == "Credit", 1)\
                  .when(col("payment_type") == "CREDIT", 1)\
                  .when(col("payment_type") == "CRD", 1)\
                  .when(col("payment_type") == "Cre", 1)\
                  .when(col("payment_type") == "CRE", 1)\
                  .when(col("payment_type") == "CASH", 2)\
                  .when(col("payment_type") == "Cash", 2)\
                  .when(col("payment_type") == "CSH", 2)\
                  .when(col("payment_type") == "CAS", 2)\
                  .when(col("payment_type") == "Cas", 2)\
                  .when(col("payment_type") == "No Charge", 3)\
                  .when(col("payment_type") == "NO CHARGE", 3)\
                  .when(col("payment_type") == "NOC", 3)\
                  .when(col("payment_type") == "NO ", 3)\
                  .when(col("payment_type") == "Dispute", 4)\
                  .when(col("payment_type") == "DISPUTE", 4)\
                  .when(col("payment_type") == "DIS", 4)\
                  .when(col("payment_type") == "Dis", 4)\
                  .when(col("payment_type") == "Unknown", 5)\
                  .when(col("payment_type") == "UNKNOWN", 5)\
                  .when(col("payment_type") == "UNK", 5)\
                  .when(col("payment_type") == "Voided Trip", 6)\
                  .when(col("payment_type") == "VOIDED TRIP", 6)\
                  .otherwise(expr("payment_type")))\
.select(\
    "year", \
    "month", \
    "pickup_datetime", \
    "dropoff_datetime", \
    "passenger_count", \
    "trip_distance", \
    "payment_type", \
    "tip_amount", \
    "trip_amount", \
    "total_amount" \
)

all_data.groupby("payment_type").count().show(20)

### Checking Schema

all_data.printSchema()

year / month should be int as well

all_data = df\
.withColumn("filename", split(input_file_name(), "/"))\
.selectExpr("*", "element_at(filename, -3) as year", "substring_index(element_at(filename, -2), '.' , 1) as month")\
.drop("filename")\
.withColumn("trip_distance", expr("trip_distance * 1.60934"))\
.withColumn("trip_amount", expr("total_amount-tip_amount"))\
.withColumn('payment_type', \
                  when(col("payment_type") == "Credit", 1)\
                  .when(col("payment_type") == "CREDIT", 1)\
                  .when(col("payment_type") == "CRD", 1)\
                  .when(col("payment_type") == "Cre", 1)\
                  .when(col("payment_type") == "CRE", 1)\
                  .when(col("payment_type") == "CASH", 2)\
                  .when(col("payment_type") == "Cash", 2)\
                  .when(col("payment_type") == "CSH", 2)\
                  .when(col("payment_type") == "CAS", 2)\
                  .when(col("payment_type") == "Cas", 2)\
                  .when(col("payment_type") == "No Charge", 3)\
                  .when(col("payment_type") == "NO CHARGE", 3)\
                  .when(col("payment_type") == "NOC", 3)\
                  .when(col("payment_type") == "NO ", 3)\
                  .when(col("payment_type") == "Dispute", 4)\
                  .when(col("payment_type") == "DISPUTE", 4)\
                  .when(col("payment_type") == "DIS", 4)\
                  .when(col("payment_type") == "Dis", 4)\
                  .when(col("payment_type") == "Unknown", 5)\
                  .when(col("payment_type") == "UNKNOWN", 5)\
                  .when(col("payment_type") == "UNK", 5)\
                  .when(col("payment_type") == "Voided Trip", 6)\
                  .when(col("payment_type") == "VOIDED TRIP", 6)\
                  .otherwise(expr("payment_type")))\
.selectExpr(\
    "cast(year as int)", \
    "cast(month as int)", \
    "pickup_datetime", \
    "dropoff_datetime", \
    "passenger_count", \
    "trip_distance", \
    "payment_type", \
    "tip_amount", \
    "trip_amount", \
    "total_amount" \
)

all_data.printSchema()

### Write results 

!hdfs dfs -rm -r /taxi/raw_all.parquet/

all_data.repartition(55).write.parquet(f"/taxi/raw_all.parquet")

### Stopping Spark 

spark.stop()
