==> V10/Jupyter_Notebooks_For_Taxi/1_Pre_Processing/5_Checks.ipynb <==
from pyspark.sql import SparkSession

spark = SparkSession \
    .builder \
    .appName("Checks") \
    .getOrCreate()

raw = spark.read.parquet("/taxi/raw_all.parquet")

import pyspark.sql.functions as f

raw.cache()
raw.count()

describe = raw.describe()

describe.show(20, False)

### Checking passenger_count, payment_type and total_amount

filtered = raw\
.filter("passenger_count < 10")\
.filter(f.expr("payment_type = 0 or payment_type = 1 or payment_type = 2 or payment_type = 3 or payment_type = 4 or payment_type = 5"))\
.filter("total_amount > 0")


filtered.count()

filtered.show()

filtered.write.parquet(f"/taxi/dataset.parquet")

### Stopping Spark 

spark.stop()
